+++ src/main/java/com/woowacamp/soolsool/core/cart/domain/CartItem.java
@@ -5,8 +5,8 @@ import static lombok.AccessLevel.PROTECTED;
 import com.woowacamp.soolsool.core.cart.domain.converter.CartItemQuantityConverter;
 import com.woowacamp.soolsool.core.cart.domain.vo.CartItemQuantity;
 import com.woowacamp.soolsool.core.liquor.domain.Liquor;
+import com.woowacamp.soolsool.global.code.GlobalErrorCode;
 import com.woowacamp.soolsool.global.common.BaseEntity;
-import com.woowacamp.soolsool.global.exception.GlobalErrorCode;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import java.util.Objects;
 import javax.persistence.Column;
@@ -51,9 +51,9 @@ public class CartItem extends BaseEntity {
 
     @Builder
     public CartItem(
-        final Long memberId,
-        final Liquor liquor,
-        final int quantity
+            final Long memberId,
+            final Liquor liquor,
+            final int quantity
     ) {
         validateIsNotNullLiquor(liquor);
