+++ src/main/java/com/woowacamp/soolsool/core/cart/domain/code/CartErrorCode.java
@@ -2,7 +2,7 @@ package com.woowacamp.soolsool.core.cart.domain.code;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
-import com.woowacamp.soolsool.global.exception.ErrorCode;
+import com.woowacamp.soolsool.global.code.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
