+++ src/main/java/com/woowacamp/soolsool/core/liquor/code/LiquorErrorCode.java
@@ -2,7 +2,7 @@ package com.woowacamp.soolsool.core.liquor.code;
 
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
-import com.woowacamp.soolsool.global.exception.ErrorCode;
+import com.woowacamp.soolsool.global.code.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
