+++ src/main/java/com/woowacamp/soolsool/core/member/code/MemberErrorCode.java
@@ -1,6 +1,6 @@
 package com.woowacamp.soolsool.core.member.code;
 
-import com.woowacamp.soolsool.global.exception.ErrorCode;
+import com.woowacamp.soolsool.global.code.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
