+++ src/main/java/com/woowacamp/soolsool/global/exception/SoolSoolException.java
@@ -1,5 +1,6 @@
 package com.woowacamp.soolsool.global.exception;
 
+import com.woowacamp.soolsool.global.code.ErrorCode;
 import lombok.Getter;
 
 @Getter
